# screen_system/openapi_before.json
{
  "components": {
    "schemas": {
      "AccIn": {
        "properties": {
          "capacity": {
            "default": 10,
            "title": "Capacity",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reorder_point": {
            "default": 5,
            "title": "Reorder Point",
            "type": "integer"
          },
          "unit": {
            "default": "箱",
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "AccIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InvAdjustIn": {
        "properties": {
          "delta": {
            "title": "Delta",
            "type": "integer"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "worker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker"
          }
        },
        "required": [
          "delta"
        ],
        "title": "InvAdjustIn",
        "type": "object"
      },
...
      "InvReorderIn": {
        "properties": {
          "reorder_point": {
            "title": "Reorder Point",
            "type": "integer"
          }
        },
        "required": [
          "reorder_point"
        ],
        "title": "InvReorderIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "スクリーン原団 工程管理 v2",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/accessories": {
      "post": {
        "operationId": "admin_accessory_create_api_admin_accessories_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Accessory Create"
      }
    },
    "/api/admin/accessories/{acc_id}": {
      "delete": {
        "operationId": "admin_accessory_delete_api_admin_accessories__acc_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "acc_id",
            "required": true,
            "schema": {
              "title": "Acc Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Accessory Delete"
      }
    },
...
    "/api/admin/inv": {
      "get": {
        "operationId": "admin_inv_list_api_admin_inv_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Inv List"
      }
    },
    "/api/admin/inv/history": {
      "get": {
        "operationId": "admin_inv_history_api_admin_inv_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 300,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Inv History"
      }
    },
    "/api/admin/inv/history/{tx_id}": {
      "delete": {
        "description": "在庫履歴を1件削除する。\n\n注意: 履歴を消しても inv_item.remain(残数)は変わらない。\n残数を直したい場合は在庫画面の「調整」を使うこと。",
        "operationId": "admin_inv_history_delete_api_admin_inv_history__tx_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "tx_id",
            "required": true,
            "schema": {
              "title": "Tx Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Inv History Delete"
      }
    },
    "/api/admin/inv/item": {
      "post": {
        "operationId": "admin_inv_item_create_api_admin_inv_item_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvItemIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Inv Item Create"
      }
    },
    "/api/admin/inv/item/{code}": {
      "delete": {
        "description": "在庫履歴を保つため論理削除(active=false)。",
        "operationId": "admin_inv_item_delete_api_admin_inv_item__code__delete",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Inv Item Delete"
      }
    },
    "/api/admin/inv/{code}/adjust": {
      "post": {
        "description": "手動補正(棚卸差異など)。シリアルは発行しない。",
        "operationId": "admin_inv_adjust_api_admin_inv__code__adjust_post",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvAdjustIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Inv Adjust"
      }
    },
    "/api/admin/inv/{code}/issue": {
      "post": {
        "description": "QR発行 = 入庫(+count)。シリアル(コード-連番)を採番して返す。",
        "operationId": "admin_inv_issue_api_admin_inv__code__issue_post",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvIssueIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Inv Issue"
      }
    },
    "/api/admin/inv/{code}/reorder": {
      "post": {
        "operationId": "admin_inv_reorder_api_admin_inv__code__reorder_post",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvReorderIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Inv Reorder"
      }
    },
    "/api/admin/inventory": {
      "get": {
        "operationId": "admin_inventory_api_admin_inventory_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Inventory"
      }
    },
    "/api/admin/inventory/adjust": {
      "post": {
        "operationId": "admin_inventory_adjust_api_admin_inventory_adjust_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Inventory Adjust"
      }
    },
    "/api/admin/inventory/history": {
      "get": {
        "operationId": "admin_inventory_history_api_admin_inventory_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "",
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "reason",
            "required": false,
            "schema": {
              "default": "",
              "title": "Reason",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "default": "",
              "title": "Date From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "default": "",
              "title": "Date To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 300,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Inventory History"
      }
    },
    "/api/admin/inventory/history.csv": {
      "get": {
        "operationId": "admin_inventory_history_csv_api_admin_inventory_history_csv_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "",
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "reason",
            "required": false,
            "schema": {
              "default": "",
              "title": "Reason",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "default": "",
              "title": "Date From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "default": "",
              "title": "Date To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "pw",
            "required": false,
            "schema": {
              "default": "",
              "title": "Pw",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Inventory History Csv"
      }
    },
    "/api/admin/inventory/reorder": {
      "post": {
        "operationId": "admin_inventory_reorder_api_admin_inventory_reorder_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-admin-pass",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Admin-Pass",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Inventory Reorder"
      }
    },
...
  }
}